backend/src/backend/requirements.txt
fastapi==0.78.0
uvicorn==0.17.6
mysql-connector-python
bcrypt